drusic/requirements.txt
discord.py
yt-dlp
python-dotenv
youtube_dl
PyNaCl
pytube
